Extract testing package
