+++ packages/testing/.mocharc.json
@@ -1,0 +1,7 @@
+{
+    "spec": "examples/**/*.test.{ts,tsx}",
+    "require": "ts-node/register",
+    "watchExtensions": "ts",
+    "extension": "ts",
+    "timeout": 3000
+}
