+++ packages/testing/examples/useBlockMeta.test.tsx
@@ -1,6 +1,6 @@
 import { expect } from 'chai'
-import { useBlockMeta } from '../src'
-import { renderWeb3Hook, sleep } from '../src/testing'
+import { useBlockMeta } from '@usedapp/core'
+import { renderWeb3Hook, sleep } from '../src'
 
 describe('useBlockMeta', () => {
   it('retrieves block timestamp and difficulty', async () => {
