+++ packages/testing/package.json
@@ -1,0 +1,58 @@
+{
+  "name": "@usedapp/testing",
+  "version": "0.1.0",
+  "main": "dist/src/index.js",
+  "types": "dist/src/index.d.ts",
+  "repository": "git@github.com:EthWorks/dapp-framework.git",
+  "author": "Ethworks",
+  "license": "MIT",
+  "devDependencies": {
+    "@ethereum-waffle/provider": "^3.2.2",
+    "@ethersproject/abi": "^5.0.12",
+    "@ethersproject/bignumber": "^5.0.14",
+    "@ethersproject/contracts": "^5.0.10",
+    "@testing-library/react-hooks": "^5.0.3",
+    "@types/chai": "^4.2.14",
+    "@types/mocha": "^8.2.0",
+    "@types/styled-components": "^5.1.6",
+    "@typescript-eslint/eslint-plugin": "^4.10.0",
+    "@typescript-eslint/parser": "^4.10.0",
+    "@web3-react/abstract-connector": "^6.0.7",
+    "@web3-react/types": "^6.0.7",
+    "chai": "^4.2.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "eslint": "7.19.0",
+    "html-webpack-plugin": "^4.5.0",
+    "jsdom": "^16.4.0",
+    "jsdom-global": "^3.0.2",
+    "mocha": "^8.2.1",
+    "prettier": "^2.1.2",
+    "ts-loader": "^8.0.11",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3",
+    "webpack": "^5.10.0",
+    "webpack-cli": "^4.2.0",
+    "webpack-dev-server": "^3.11.0"
+  },
+  "dependencies": {
+    "@ethersproject/providers": "^5.0.17",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "@usedapp/core": "*",
+    "@web3-react/core": "^6.1.1",
+    "@web3-react/injected-connector": "^6.0.7",
+    "ethereum-waffle": "^3.2.2",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "styled-components": "^5.2.1"
+  },
+  "scripts": {
+    "start": "webpack serve --mode development",
+    "build": "tsc",
+    "test": "mocha -r jsdom-global/register",
+    "lint": "yarn lint:prettier --check && yarn lint:eslint",
+    "lint:fix": "yarn lint:prettier --write && yarn lint:eslint --fix",
+    "lint:eslint": "eslint './{src, test}/**/*.{ts,tsx}'",
+    "lint:prettier": "yarn prettier './{src, test}/**/*.{ts,tsx}'"
+  }
+}
